+++ src/tagfilter.rs
@@ -137,9 +137,111 @@ impl std::str::FromStr for TagFilter {
     }
 }
 
+#[derive(Debug, Clone, PartialEq)]
+pub enum TagFilterFuncElement {
+    FilterMatchThenTrue(TagFilter),
+    FilterMatchThenFalse(TagFilter),
+    AlwaysTrue,
+    AlwaysFalse,
+}
+
+impl std::str::FromStr for TagFilterFuncElement {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let s = s.trim();
+        if s == "T" {
+            Ok(TagFilterFuncElement::AlwaysTrue)
+        } else if s == "F" {
+            Ok(TagFilterFuncElement::AlwaysFalse)
+        } else if let Some(filter) = s.strip_suffix("→T") {
+            let filter: TagFilter = filter.parse()?;
+            Ok(TagFilterFuncElement::FilterMatchThenTrue(filter))
+        } else if let Some(filter) = s.strip_suffix("→F") {
+            let filter: TagFilter = filter.parse()?;
+            Ok(TagFilterFuncElement::FilterMatchThenFalse(filter))
+        } else {
+            Err(format!("Unknown Tag Filter Func: {}", s))
+        }
+    }
+}
+
+impl ToString for TagFilterFuncElement {
+    fn to_string(&self) -> String {
+        match self {
+            TagFilterFuncElement::AlwaysTrue => "T".to_string(),
+            TagFilterFuncElement::AlwaysFalse => "F".to_string(),
+            TagFilterFuncElement::FilterMatchThenTrue(f) => format!("{}→T", f.to_string()),
+            TagFilterFuncElement::FilterMatchThenFalse(f) => format!("{}→F", f.to_string()),
+        }
+    }
+}
+
+impl TagFilterFuncElement {
+    pub fn result(&self, o: &impl osmio::OSMObjBase) -> Option<bool> {
+        match self {
+            TagFilterFuncElement::AlwaysTrue => Some(true),
+            TagFilterFuncElement::AlwaysFalse => Some(false),
+            TagFilterFuncElement::FilterMatchThenTrue(f) => {
+                if f.filter(o) {
+                    Some(true)
+                } else {
+                    None
+                }
+            }
+            TagFilterFuncElement::FilterMatchThenFalse(f) => {
+                if f.filter(o) {
+                    Some(false)
+                } else {
+                    None
+                }
+            }
+        }
+    }
+}
+
+// waterway=canal∧lock=yes→T;waterway=canal→F;waterway→T;F
+#[derive(Debug, Clone, PartialEq)]
+pub struct TagFilterFunc(Vec<TagFilterFuncElement>);
+
+impl std::str::FromStr for TagFilterFunc {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let s = s.trim();
+        let tffs = s
+            .split(';')
+            .map(|tff| tff.parse::<TagFilterFuncElement>())
+            .collect::<Result<Vec<_>, _>>()?;
+        Ok(TagFilterFunc(tffs))
+    }
+}
+
+impl ToString for TagFilterFunc {
+    fn to_string(&self) -> String {
+        let parts = self
+            .0
+            .iter()
+            .map(|tff| tff.to_string())
+            .collect::<Vec<String>>();
+        parts.join(";")
+    }
+}
+
+impl TagFilterFunc {
+    pub fn result(&self, o: &impl osmio::OSMObjBase) -> Option<bool> {
+        self.0
+            .iter()
+            .map(|tff| tff.result(o))
+            .find(|res| res.is_some())
+            .flatten()
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use osmio::OSMObjBase;
 
     macro_rules! test_parse {
         ( $name:ident, $input:expr, $expected_output:expr ) => {
@@ -252,4 +354,142 @@ mod tests {
             ])
         );
     }
+
+    #[test]
+    fn tag_filter_func_parse() {
+        assert_eq!(
+            "T".parse::<TagFilterFuncElement>().unwrap(),
+            TagFilterFuncElement::AlwaysTrue
+        );
+        assert_eq!(
+            "F".parse::<TagFilterFuncElement>().unwrap(),
+            TagFilterFuncElement::AlwaysFalse
+        );
+        assert_eq!(
+            "highway→T".parse::<TagFilterFuncElement>().unwrap(),
+            TagFilterFuncElement::FilterMatchThenTrue(TagFilter::HasK("highway".to_string()))
+        );
+        assert_eq!(
+            "waterway→F".parse::<TagFilterFuncElement>().unwrap(),
+            TagFilterFuncElement::FilterMatchThenFalse(TagFilter::HasK("waterway".to_string()))
+        );
+        assert_eq!(
+            "waterway=canal→F".parse::<TagFilterFuncElement>().unwrap(),
+            TagFilterFuncElement::FilterMatchThenFalse(TagFilter::KV(
+                "waterway".to_string(),
+                "canal".to_string()
+            ))
+        );
+        assert_eq!(
+            "waterway=canal∧lock=yes→F"
+                .parse::<TagFilterFuncElement>()
+                .unwrap(),
+            TagFilterFuncElement::FilterMatchThenFalse(TagFilter::And(vec![
+                TagFilter::KV("waterway".to_string(), "canal".to_string()),
+                TagFilter::KV("lock".to_string(), "yes".to_string())
+            ]))
+        );
+        assert_eq!(
+            "waterway=canal∧usage∈headrace,tailrace→F"
+                .parse::<TagFilterFuncElement>()
+                .unwrap(),
+            TagFilterFuncElement::FilterMatchThenFalse(TagFilter::And(vec![
+                TagFilter::KV("waterway".to_string(), "canal".to_string()),
+                TagFilter::KinV(
+                    "usage".to_string(),
+                    vec!["headrace".to_string(), "tailrace".to_string()]
+                )
+            ]))
+        );
+
+        assert!("highway".parse::<TagFilterFuncElement>().is_err());
+        assert!("highway=primary".parse::<TagFilterFuncElement>().is_err());
+    }
+
+    macro_rules! test_parse_tag_filter_func {
+        ( $name:ident, $input_tff:expr, $input_tags:expr, $expected_output:expr ) => {
+            #[test]
+            fn $name() {
+                let tff: TagFilterFuncElement = $input_tff.parse().unwrap();
+                let mut n = osmio::obj_types::StringNodeBuilder::default()
+                    ._id(1)
+                    .build()
+                    .unwrap();
+                for (k, v) in $input_tags.iter() {
+                    n.set_tag(k.to_string(), v.to_string());
+                }
+                assert_eq!(tff.result(&n), $expected_output);
+            }
+        };
+    }
+    test_parse_tag_filter_func!(tff_parse1, "T", [("highway", "yes")], Some(true));
+    test_parse_tag_filter_func!(tff_parse2, "F", [("highway", "yes")], Some(false));
+    test_parse_tag_filter_func!(tff_parse3, "highway→T", [("highway", "yes")], Some(true));
+    test_parse_tag_filter_func!(tff_parse4, "highway→T", [("natural", "water")], None);
+    test_parse_tag_filter_func!(
+        tff_parse5,
+        "natural=water→T",
+        [("natural", "water")],
+        Some(true)
+    );
+    test_parse_tag_filter_func!(tff_parse6, "natural=water→T", [("natural", "rock")], None);
+
+    #[test]
+    fn tag_filter_func_list_parse() {
+        assert_eq!(
+            "T".parse::<TagFilterFunc>().unwrap(),
+            TagFilterFunc(vec![TagFilterFuncElement::AlwaysTrue])
+        );
+        assert_eq!(
+            "waterway→T;F".parse::<TagFilterFunc>().unwrap(),
+            TagFilterFunc(vec![
+                TagFilterFuncElement::FilterMatchThenTrue(TagFilter::HasK("waterway".to_string())),
+                TagFilterFuncElement::AlwaysFalse
+            ])
+        );
+    }
+
+    macro_rules! test_parse_tag_filter_func_list {
+        ( $name:ident, $input_tff:expr, $input_tags:expr, $expected_output:expr ) => {
+            #[test]
+            fn $name() {
+                let tff: TagFilterFunc = $input_tff.parse().unwrap();
+                let mut n = osmio::obj_types::StringNodeBuilder::default()
+                    ._id(1)
+                    .build()
+                    .unwrap();
+                for (k, v) in $input_tags.iter() {
+                    n.set_tag(k.to_string(), v.to_string());
+                }
+                assert_eq!(tff.result(&n), $expected_output);
+            }
+        };
+    }
+
+    test_parse_tag_filter_func_list!(tffl_parse1, "T", [("highway", "yes")], Some(true));
+    test_parse_tag_filter_func_list!(tffl_parse2, "highway→T;F", [("highway", "yes")], Some(true));
+    test_parse_tag_filter_func_list!(
+        tffl_parse3,
+        "highway→T;F",
+        [("natural", "yes")],
+        Some(false)
+    );
+    test_parse_tag_filter_func_list!(
+        tffl_parse4,
+        "waterway=canal→F;waterway→T;F",
+        [("waterway", "river")],
+        Some(true)
+    );
+    test_parse_tag_filter_func_list!(
+        tffl_parse5,
+        "waterway=canal→F;waterway→T;F",
+        [("highway", "primary")],
+        Some(false)
+    );
+    test_parse_tag_filter_func_list!(
+        tffl_parse6,
+        "waterway=canal→F;waterway→T;F",
+        [("waterway", "canal")],
+        Some(false)
+    );
 }