+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * AND'ing filters together now possible with `∧` (`U+2227 LOGICAL AND`):
+* New “Tag Filter Function” allows much more powerful, and complicated, tag
+  filtering
 
 ## v0.20.0 (2024-02-28)